Match full variety of Content-Type values
At the moment the "Content-Type" header must be exactly 'application/json' for the refresh token to be extracted from the request. 

It is valid for "Content-Type" header to have a parameter for example 'application/json;charset=UTF-8', but this does not match the criteria and so the refresh fails. 

This change checks that the Content-Type header starts with  'application/json' and will then attempt to extract refresh_token from the content of the request.