Fixes #287 - Add LogoutEventListener
This commit introduces a `LogoutEventListener` which invalidates the
given `refresh_token` and unsets the cookie, if enabled.

If no refresh token is supplied, an error is returned and the cookie
remains untouched. If the supplied refresh token is (already) invalid,
the cookie is unset.

Because the `LogoutEventListener` always sets a response, it would
inhibit normal logout behavior and therefore should only run on a
specifically configured firewall.

Therefore a new configuration option is introduced, called
`logout_firewall`, which contains the name of the firewall that
triggers the logout event we want to hook into (default: `api`).